api/requirements.txt
diff-match-patch==20121119
Django==1.9.4
django-floppyforms==1.6.1
django-import-export==0.4.3
djangorestframework==3.3.2
Jinja2==2.8
MarkupSafe==0.23
tablib==0.11.2
